+++ lang/en/local_boostnavigation.php
@@ -115,14 +115,14 @@ Please note:
 <li>If the custom node does not show up in Boost\'s nav drawer, please check if all mandatory params are set correctly and if the optional language setting fits to your current Moodle user language.</li>
 <li>Due to the way how Boost\'s nav drawer is built in Moodle core, all custom nodes are displayed equally. Adding custom CSS classes, custom HTML element ids or a target attribute to open the link in a new window is impossible.</li>
 </ul>';
-$string['setting_customnodesusageusers'] = 'Each line consists of a link title, a link URL, supported language(s) (optional), supported cohort(s) (optional), supported role(s) (optional), supported global roles(s) (optional), the logical combination operator (optional), an icon (optional) and the ID attribute (optional) - separated by pipe characters. Each custom node needs to be written in a new line.<br />
+$string['setting_customnodesusageusers'] = 'Each line consists of a link title, a link URL, supported language(s) (optional), supported cohort(s) (optional), supported role(s) (optional), supported global roles(s) (optional), the logical combination operator (optional), an icon (optional), the ID attribute (optional) and a \'before node key\' (optional) - separated by pipe characters. Each custom node needs to be written in a new line.<br />
 For example:<br />
 Moodle.org website|http://www.moodle.org|en,de<br />
 Our university|http://www.our-university.edu<br />
 Faculty of mathematics|http://www.our-university.edu/math||math<br />
 Teachers\' handbook|http://www.our-university.edu/teacher-handbook|||editingteacher,teacher<br />
 Student information course|/course/view.php?id=1234||||||fa-graduation-cap<br />
-{editingtoggle}|/course/view.php?id={courseid}&sesskey={sesskey}&edit={editingtoggle}|||editingteacher|admin,manager|OR|fa-pencil|editing<br /><br />
+{editingtoggle}|/course/view.php?id={courseid}&sesskey={sesskey}&edit={editingtoggle}|||editingteacher|admin,manager|OR|fa-pencil|editing|participants<br /><br />
 Further information to the parameters:
 <ul>
 <li><b>Title:</b> This text will be shown as the clickable text / label of the custom node.</li>
@@ -134,6 +134,7 @@ Further information to the parameters:
 <li><b>Logical combination operator (optional):</b> This setting can be used to change the logical combination of the optional parameters for cohort, role and system role. If you set this setting to AND or leave this setting empty, the optional parameters for cohort, role and system role will be logically combined with AND and the custom node will only be shown if ALL parameters apply. If you set this setting to OR, the optional parameters for cohort, role and system role will be logically combined with OR and the custom node will be shown if ANY of the parameters apply. This gives you the flexibility to create, for example, a node which is shown to users with a given course role as well as to users with a given system role as shown in our example above.</li>
 <li><b>Icon (optional):</b> This icon will be used as icon for the custom node, for example fa-flag. Use a Font Awesome icon identifier (<a href="https://fontawesome.com/v4.7.0/icons/">See the icon list on fontawesome.com</a>) to identify which icon should be used. Font Awesome is included in Boost, classic Moodle pix icons are not supported here. If you just want to use a standard bullet icon for the custom node, then leave this field empty. Custom nodes which don\'t get a standard bullet icon automatically also can\'t get a custom icon currently.</li>
 <li><b>ID attribute (optional):</b> This setting can be used for setting the ID attribute of the custom node to the given fixed value. The given ID value is cleaned for alphanumeric characters only and prefixed by the particular custom node area, for example setting the ID of a node in the bottom nodes for admins area to \'my_node\' will result in an ID attribute filled with \'localboostnavigationcustombottomadminsmynode\'. This setting gives you the ability to reference a custom node in CSS selectors or even in custom JavaScript code. Setting the same ID to multiple nodes is not forbidden here, but please note that duplicate ID attributes are discouraged in HTML pages. If this field is empty, the custom node will get an automatically generated ID which is not really suitable for referencing.</li>
+<li><b>Before node key (optional):</b> This setting can be used to specify the node key above which the custom node should be placed. You find the node key in the \'data-key\' HTML attribute of the node you want to reference here. In case of a set of nodes made of a parent and one or more children, this setting must be given to each node of the set. Please note that this setting is considered partly experimental as not all nodes can be addressed with this setting yet.</li>
 </ul>
 Please note:
 <ul>